Finished hide/show functionality for user login
