+++ scripts/main.js
@@ -1,1 +1,4 @@
-const registerOrLogin = require("./registerVerify")
+const registerOrLogin = require("./register_login/registerVerify")
+const pageShown = require("./register_login/dashboard_hide_show")
+
+pageShown()
