+++ scripts/register_login/dashboard_hide_show.js
@@ -1,0 +1,13 @@
+// This hides/shows the login or dashboard based on whether the user is "active"
+function pageShown() {
+    const showCurrentUserPage = JSON.parse(sessionStorage.getItem("activeUser"))
+    console.log("In pageShown")
+    if (showCurrentUserPage !== null) {
+        document.getElementById("loginPage").classList.add("hidden")
+        document.getElementById("dashboardPage").classList.remove("hidden")
+    } else {
+        document.getElementById("dashboardPage").classList.add("hidden")
+    }
+}
+
+module.exports = pageShown
